chatbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
langchain-core>=0.1.10
langchain-astradb>=0.1.1
langchain-huggingface>=0.0.1
pypdf2==3.0.1
sentence-transformers==2.6.1
faiss-cpu
python-dotenv==1.0.1
googletrans==4.0.0rc1
openai>=1.0.0
pydantic>=2.0.0
